backend/config.py
from pathlib import Path

from dotenv import load_dotenv

ROOT_DIR = Path(__file__).resolve().parent.parent
ENV_FILE = ROOT_DIR / ".env"


def load_local_env() -> None:
    load_dotenv(ENV_FILE, override=False)
