pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deckly"
version = "0.1.0"
description = "Flashcard decks and quiz workflows served by FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "SQLAlchemy==2.0.28",
    "bcrypt==4.1.2",
    "python-jose==3.3.0",
    "pydantic[email]==1.10.14",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-cov==5.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]
